+++ CHANGELOG.md
@@ -24,6 +24,7 @@ FlatLaf Change Log
   another screen with different scaling factor. (issue #166)
 - Fixed occasional wrong positioning of heavy weight popups when using multiple
   screens with different scaling factors. (issue #166)
+- ToolTip: Avoid that tooltip hides owner component. (issue #164)
 
 
 ## 0.41